merge: merge main code and add AuthPosts logic

+++ src/main/java/B1ND/linkUp/domain/post/service/PostsCommentService.java
@@ -1,5 +1,6 @@
 package B1ND.linkUp.domain.post.service;
 
+import B1ND.linkUp.domain.auth.entity.User;
 import B1ND.linkUp.domain.post.dto.request.PostsCommentRequest;
 import B1ND.linkUp.domain.post.dto.response.MessageResponse;
 import B1ND.linkUp.domain.post.entity.Posts;
@@ -9,6 +10,7 @@ import B1ND.linkUp.domain.post.exception.PostsException;
 import B1ND.linkUp.domain.post.repository.PostsCommentRepository;
 import B1ND.linkUp.domain.post.repository.PostsRepository;
 import B1ND.linkUp.global.common.APIResponse;
+import B1ND.linkUp.global.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.aspectj.bridge.Message;
 import org.springframework.stereotype.Service;
@@ -18,14 +20,18 @@ import org.springframework.stereotype.Service;
 public class PostsCommentService {
     private final PostsRepository postsRepository;
     private final PostsCommentRepository postsCommentRepository;
+    private final SecurityUtil securityUtil;
 
     public APIResponse<MessageResponse> createComment(Long id, PostsCommentRequest request) {
         Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new PostsException(PostsErrorCode.POST_NOT_FOUND));
 
+        User user = securityUtil.getUser();
+
         postsCommentRepository.save(PostsComment.builder()
                 .content(request.content())
                 .posts(posts)
+                .user(user)
                 .build());
         return APIResponse.ok(MessageResponse.of("답변이 등록되었습니다."));
     }
@@ -34,6 +40,11 @@ public class PostsCommentService {
         PostsComment comment = postsCommentRepository.findById(id)
                 .orElseThrow(() -> new PostsException(PostsErrorCode.ANSWER_NOT_FOUND));
 
+        User user = securityUtil.getUser();
+
+        if(!user.equals(comment.getUser())) {
+            throw new PostsException(PostsErrorCode.NOT_POST_AUTHOR);
+        }
         postsCommentRepository.delete(comment);
 
         return APIResponse.ok(MessageResponse.of("답변이 삭제되었습니다."));
@@ -42,9 +53,15 @@ public class PostsCommentService {
     public APIResponse<MessageResponse> AnswerAcceptanceService(Long id, Long answerId) {
         Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new PostsException(PostsErrorCode.POST_NOT_FOUND));
+        User user = securityUtil.getUser();
+
         if (posts.isAccepted()) {
             throw new PostsException(PostsErrorCode.ALREADY_ACCEPTED_ANSWER);
         }
+        if (!user.equals(posts.getUser())) {
+            throw new PostsException(PostsErrorCode.NOT_POST_AUTHOR);
+        }
+
         PostsComment comment = postsCommentRepository.findByIdAndPosts(answerId, posts)
                 .orElseThrow(() -> new PostsException(PostsErrorCode.ANSWER_NOT_FOUND));
