+++ src/main/java/B1ND/linkUp/domain/post/service/PostsService.java
@@ -1,5 +1,6 @@
 package B1ND.linkUp.domain.post.service;
 
+import B1ND.linkUp.domain.auth.entity.User;
 import B1ND.linkUp.domain.post.dto.request.CreatePostsRequest;
 import B1ND.linkUp.domain.post.dto.request.ReadPostsRequest;
 import B1ND.linkUp.domain.post.dto.request.UpdatePostsRequest;
@@ -13,13 +14,13 @@ import B1ND.linkUp.domain.post.repository.PostsLikeRepository;
 import B1ND.linkUp.domain.post.repository.PostsRepository;
 import B1ND.linkUp.global.common.APIResponse;
 import B1ND.linkUp.global.common.PageResponse;
+import B1ND.linkUp.global.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,6 +30,7 @@ import java.util.List;
 public class PostsService {
     private final PostsRepository postsRepository;
     private final PostsLikeRepository postsLikeRepository;
+    private final SecurityUtil securityUtil;
 
     public APIResponse<PageResponse<List<ReadPostsResponse>>> ReadPosts(int page, ReadPostsRequest req) {
         Pageable pageable = PageRequest.of(page, 10, Sort.by(Sort.Direction.DESC, "id"));
@@ -47,18 +49,22 @@ public class PostsService {
         Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new PostsException(PostsErrorCode.POST_NOT_FOUND));
 
-        boolean isLike = postsLikeRepository.existsByPosts_IdAndMember_Id(posts.getId(),1L); //memberId같은 경우 머지된 후 추가할 예정
+        User user = securityUtil.getUser();
+
+        boolean isLike = postsLikeRepository.existsByPosts_IdAndUser(posts.getId(),user);
 
         return APIResponse.ok(ViewPostsResponse.of(posts, isLike));
     }
 
     public APIResponse<MessageResponse> createPosts(CreatePostsRequest req) {
+        User user = securityUtil.getUser();
         postsRepository.save(
-                        Posts.builder()
+                Posts.builder()
                         .title(req.title())
                         .content(req.content())
                         .author(req.author())
                         .category(req.category())
+                        .user(user)
                         .build()
         );
         return APIResponse.of(HttpStatus.CREATED, MessageResponse.of("게시글이 등록되었습니다."));
@@ -68,6 +74,11 @@ public class PostsService {
         Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new PostsException(PostsErrorCode.POST_NOT_FOUND));
 
+        User user = securityUtil.getUser();
+        if(!user.equals(posts.getUser())) {
+            throw new PostsException(PostsErrorCode.NOT_POST_AUTHOR);
+        }
+
         posts.updatePosts(request);
         postsRepository.save(posts);
         return APIResponse.ok(MessageResponse.of("게시글이 수정되었습니다."));
@@ -76,6 +87,11 @@ public class PostsService {
     public APIResponse<MessageResponse> deletePosts(Long id) {
         Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new PostsException(PostsErrorCode.POST_NOT_FOUND));
+
+        User user = securityUtil.getUser();
+        if(!user.equals(posts.getUser())) {
+            throw new PostsException(PostsErrorCode.NOT_POST_AUTHOR);
+        }
 
         postsRepository.delete(posts);
         return APIResponse.ok(MessageResponse.of("게시글이 삭제되었습니다."));
